Horizontal Green-Ampt flow integrated

This function models saturated horizontal flow from a saturated surface into a soil profile. The assumptions are the same as in the Green-Ampt equation, except that there is no effect of gravity because all flow is assumed to be horizontal. The function further integrates over the vertical depth of a pit such that the hydraulic head at the boundary goes from (h_b - thickness) to h_b. The equation is:

$$F^2$$ $$F^2$$ $$F^2$$

get_greenampt_horiz_flow_integrated(VWC_0, n, Ksat, h_b, h_0, t, d = NULL)

Arguments

VWC_0

Soil volumetric water content prior to event

n

Soil porosity

Ksat

Saturated hydraulic conductivity

h_b

Hydraulic head at soil surface boundary

h_0

Hydraulic head in soil prior to event

thickness

Depth over which to be integrated. If NULL, set to h_b

Value

Returns the time at which a cumulative amount of infiltration occurs.

Examples

library(units)
VWC_0 <- 0.2 # unitless
n <- 0.35 # unitless
Ksat <- set_units(0.2, "cm/h") # length / time
h_b <- set_units(6, "ft") # hydraulic head (length)
h_0 <- set_units(-10, "cm") # hydraulic head (length)

Fv <- get_greenampt_horiz_flow_integrated(VWC_0, n, Ksat, h_b, h_0, t = set_units(1,"hr"), d = NULL)

# Get the infiltration over a 1 mm differential depth to compare with the point infiltration
Fv_dv <- get_greenampt_horiz_flow_integrated(VWC_0, n, Ksat, h_b, h_0, t = set_units(1,"hr"), d = set_units(1, "mm"))
Fv_point <- Fv_dv/ set_units(1, "mm")
Fv_point
#> 34.01441 [mm]
# Get the point infiltration
F_point <- get_greenampt_horiz_flow(VWC_0, n, Ksat, h_b, h_0, t = set_units(1,"hr")) %>% set_units("mm")
F_point
#> 34.01882 [mm]
# percent error:
(Fv_point - F_point) / F_point * 100
#> -0.01296255 [1]